Made Lotus class, start of B-Spline class, fixed uint
Completed part 3 of the project, hard coding a curve that forms
a shape vaguely reminiscent of a lotus blossom. Also created the
skeleton for the B-Spline class to be implemented next.

As always, several uses of 'uint' in GMlib had to be corrected,
such instances were found and replaced with 'unsigned int'