BlendingSpline can now mostly copy other curves
Still not completely satisfied with the Blending Spline on the
matter of copying the Lotus model curve. A simpler model curve
might be necessary. The Blending Spline struggles on the parts
of the lotus with a smaller radius. It might also be related to
the weld in a closed curve.