Made some progress on BlendingSurface, mostly fixed GMlib problems
By some strange fault, GMlib was only compiled once on my laptop PC,
and any modifications never compiled here. This made it seem like
changes that do not work on the desktop PC works on the laptop.
The changes made to GMlib should hopefully enable the desktop PC to
run the program, as the error was recreated and fixed in this commit.